pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = ["Roman Volodin <rozarioagro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
flask-limiter = "^3.5.1"
flask-swagger-ui = "^4.11.1"
sqlalchemy = "^2.0.28"
flask-jwt-extended = "^4.6.0"
flask-marshmallow = "^1.2.0"
flask-migrate = "^4.0.6"
flask-mobility = "^2.0.1"
flask-caching = "^2.1.0"
redis = "^5.0.3"
marshmallow-enum = "^1.5.1"
marshmallow-sqlalchemy = "^1.0.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
